Introduce IndexedKeySet
This solves having a map being passed in several places, making
the code more readable and it also removes code duplication
for generating the final output from the keyset.

Switched to SHA256, as MD5 has been deprecated for any kind
of use a long time ago. (and the overhead with current cpu's
is so minimal that we really will not notice it).